+++ requirements.txt
@@ -1,1 +1,1 @@
-discord.py==1.3.3
+discord.py==1.5.1
